+++ styles/components/_ModalContent.scss
