+++ styles/styles.scss
@@ -10,7 +10,7 @@
 @import './components/ModeRules';
 @import './components/Buttons';
 @import './components/LightsaberTimer';
-@import './components/EndGame';
+@import './components/ModalContent';
 
 body {
   font-family: 'Montserrat', sans-serif;
